+++ Sources/TuistEnvKit/Installer/Installer.swift
@@ -181,6 +181,12 @@ final class Installer: Installing {
             }
             try FileHandler.shared.createFolder(installationDirectory)
 
+            let templatesDirectory = temporaryDirectory.path.appending(component: Constants.templatesDirectoryName)
+            if FileHandler.shared.exists(templatesDirectory) {
+                try FileHandler.shared.copy(from: templatesDirectory,
+                                            to: buildDirectory.appending(component: Constants.templatesDirectoryName))
+            }
+
             try buildCopier.copy(from: buildDirectory,
                                  to: installationDirectory)