+++ Sources/TuistScaffold/TemplatesDirectoryLocator.swift
@@ -4,6 +4,8 @@ import TuistCore
 import TuistSupport
 
 public protocol TemplatesDirectoryLocating {
+    /// Returns the path to the tuist built-in templates directory if it exists.
+    func locateTuistTemplates() -> AbsolutePath?
     /// Returns the path to the user-defined templates directory if it exists.
     /// - Parameter at: Path from which we traverse the hierarchy to obtain the templates directory.
     func locateUserTemplates(at: AbsolutePath) -> AbsolutePath?
@@ -21,6 +23,27 @@ public final class TemplatesDirectoryLocator: TemplatesDirectoryLocating {
 
     // MARK: - TemplatesDirectoryLocating
 
+    public func locateTuistTemplates() -> AbsolutePath? {
+        #if DEBUG
+            // Used only for debug purposed to find templates in your tuist working directory
+            // `bundlePath` points to tuist/Templates
+            let bundlePath = AbsolutePath(#file.replacingOccurrences(of: "file://", with: ""))
+                .removingLastComponent()
+                .removingLastComponent()
+                .removingLastComponent()
+        #else
+            let bundlePath = AbsolutePath(Bundle(for: TemplatesDirectoryLocator.self).bundleURL.path)
+        #endif
+        let paths = [
+            bundlePath,
+            bundlePath.parentDirectory,
+        ]
+        let candidates = paths.map { path in
+            path.appending(component: Constants.templatesDirectoryName)
+        }
+        return candidates.first(where: FileHandler.shared.exists)
+    }
+
     public func locateUserTemplates(at: AbsolutePath) -> AbsolutePath? {
         guard let customTemplatesDirectory = locate(from: at) else { return nil }
         if !FileHandler.shared.exists(customTemplatesDirectory) { return nil }
@@ -28,9 +51,11 @@ public final class TemplatesDirectoryLocator: TemplatesDirectoryLocating {
     }
 
     public func templateDirectories(at path: AbsolutePath) throws -> [AbsolutePath] {
-        let customTemplatesDirectory = locateUserTemplates(at: path)
-        let customTemplates = try customTemplatesDirectory.map(FileHandler.shared.contentsOfDirectory) ?? []
-        return customTemplates.filter(FileHandler.shared.isFolder)
+        let tuistTemplatesDirectory = locateTuistTemplates()
+        let tuistTemplates = try tuistTemplatesDirectory.map(FileHandler.shared.contentsOfDirectory) ?? []
+        let userTemplatesDirectory = locateUserTemplates(at: path)
+        let userTemplates = try userTemplatesDirectory.map(FileHandler.shared.contentsOfDirectory) ?? []
+        return (tuistTemplates + userTemplates).filter(FileHandler.shared.isFolder)
     }
 
     // MARK: - Helpers